kvserver: move raft log storage to a file
This commit mechanically moves the implementation of the raft LogStorage
interface into a new file.

Epic: none
Release note: none
