+++ pkg/kv/kvserver/BUILD.bazel
@@ -65,6 +65,7 @@ go_library(
         "replica_raft.go",
         "replica_raft_overload.go",
         "replica_raft_quiesce.go",
+        "replica_raftlog.go",
         "replica_raftstorage.go",
         "replica_range_lease.go",
         "replica_rangefeed.go",
